Changed Affiliate Model to be consistent for Affiliate and Program Endpoints
