+++ src/Model/ProgramAffiliate.cs
@@ -1,0 +1,15 @@
+﻿using Newtonsoft.Json;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TapfiliateNet.Model
+{
+    public class ProgramAffiliate : AffiliateMaster
+    {
+        [JsonProperty("referral_link")]
+        public string ReferralLink { get; set; }
+    }
+}
