+++ src/TapfiliateNet.csproj
@@ -47,14 +47,17 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Model\Affiliate.cs" />
+    <Compile Include="Model\AffiliateMaster.cs" />
     <Compile Include="Model\Commission.cs" />
     <Compile Include="Model\Conversion.cs" />
     <Compile Include="Model\Payout.cs" />
     <Compile Include="Model\Program.cs" />
+    <Compile Include="Model\ProgramAffiliate.cs" />
     <Compile Include="Request\AffiliateRequest.cs" />
     <Compile Include="Request\CommissionRequest.cs" />
     <Compile Include="Request\CommissionUpdateRequest.cs" />
     <Compile Include="Request\ConversionRequest.cs" />
+    <Compile Include="Request\ProgramAddAffiliateRequest.cs" />
     <Compile Include="TapfiliateClient.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
